requirements.txt
# Core Python utilities
numpy>=1.24.0
matplotlib>=3.7.0
tqdm>=4.65.0
Pillow>=10.0.0
glob2>=0.7

# Image processing
scikit-image>=0.21.0
scipy>=1.11.0

# FastAI (for dataset helpers)
fastai>=2.7.14

# Optional utilities
pathlib2>=2.3.7.post1  # (included in stdlib but safe for back-compat)

# Model/runtime dependencies
torch>=2.1.0
torchvision>=0.18.0

# Web tier
fastapi>=0.115.0
uvicorn[standard]>=0.23.2
python-multipart>=0.0.6
aiofiles>=23.2.1
opencv-python-headless>=4.9.0.74
